+++ benches/mvg.rs
@@ -12,7 +12,7 @@ fn bench_mvg_draw(c: &mut Criterion) {
         group.bench_with_input(format!("{dims} dims"), &dims, |b, &dims| {
             let mvg = MvGaussian::standard(dims).unwrap();
             b.iter_batched_ref(
-                rand::thread_rng,
+                rand::rng,
                 |mut rng| black_box::<DVector<f64>>(mvg.draw(&mut rng)),
                 BatchSize::SmallInput,
             );
@@ -28,7 +28,7 @@ fn bench_mvg_sample(c: &mut Criterion) {
     for n in [1, 10, 50, 100] {
         group.bench_with_input(n.to_string(), &n, |b, &n| {
             b.iter_batched_ref(
-                rand::thread_rng,
+                rand::rng,
                 |mut rng| {
                     black_box::<Vec<DVector<f64>>>(mvg.sample(n, &mut rng))
                 },