+++ examples/coin_flips.rs
@@ -1,18 +1,18 @@
 use rand::Rng;
+use rv::ConjugateModel;
 use rv::dist::{Bernoulli, Beta};
 use rv::prelude::BernoulliData;
 use rv::traits::{ConjugatePrior, DataOrSuffStat, Mean, Sampleable, SuffStat};
-use rv::ConjugateModel;
 use std::sync::Arc;
 
 fn main() {
-    let mut rng = rand::thread_rng();
+    let mut rng = rand::rng();
 
     // Generate some 1000 coin flips from a coin that will come up head 70%
     // of the time.
     let flips: Vec<bool> = (0..1000)
         .map(|_| {
-            let x: f64 = rng.gen();
+            let x: f64 = rng.random();
             x < 0.7
         })
         .collect();