+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 
 ### Changed
-- Merry Christmas
+- Updated to Rust 2024 Edition
+- Updated `rand`, `rand_xoshiro`, and `rand_distr` deps.
 - `NormalInvChiSquared`, `NormalGamma`, and `NormalInvGamme` `PpCache` for Gaussian conjugate analysis changed. `ln_pp_with_cache` is much faster.
 - `Gamma` `PpCache` for Poisson conjugate analysis has been optimized. `ln_pp_with_cache` is faster.
 - Fixed some typos.